+++ lib/phoenix_kit_web/components/admin_nav.ex
@@ -145,6 +145,8 @@ defmodule PhoenixKitWeb.Components.AdminNav do
           <.icon name="hero-map" class="w-5 h-5" />
         <% "document" -> %>
           <.icon name="hero-document-text" class="w-5 h-5" />
+        <% "legal" -> %>
+          <.icon name="hero-scale" class="w-5 h-5" />
         <% "maintenance" -> %>
           <.icon name="hero-wrench-screwdriver" class="w-5 h-5" />
         <% "storage" -> %>