+++ lib/phoenix_kit_web/components/layout_wrapper.ex
@@ -744,6 +744,7 @@ defmodule PhoenixKitWeb.Components.LayoutWrapper do
                           "/admin/settings/storage/dimensions",
                           "/admin/settings/maintenance",
                           "/admin/settings/blogging",
+                          "/admin/settings/legal",
                           "/admin/settings/seo",
                           "/admin/settings/posts",
                           "/admin/settings/billing",
@@ -752,7 +753,7 @@ defmodule PhoenixKitWeb.Components.LayoutWrapper do
                       }
                     />
 
-                    <%= if submenu_open?(@current_path, ["/admin/settings", "/admin/settings/users", "/admin/settings/referral-codes", "/admin/settings/emails", "/admin/settings/languages", "/admin/settings/entities", "/admin/settings/media", "/admin/settings/storage/dimensions", "/admin/settings/maintenance", "/admin/settings/blogging", "/admin/settings/seo", "/admin/settings/sitemap", "/admin/settings/posts", "/admin/settings/billing", "/admin/settings/billing/providers"]) do %>
+                    <%= if submenu_open?(@current_path, ["/admin/settings", "/admin/settings/users", "/admin/settings/referral-codes", "/admin/settings/emails", "/admin/settings/languages", "/admin/settings/entities", "/admin/settings/media", "/admin/settings/storage/dimensions", "/admin/settings/maintenance", "/admin/settings/blogging", "/admin/settings/legal", "/admin/settings/seo", "/admin/settings/sitemap", "/admin/settings/posts", "/admin/settings/billing", "/admin/settings/billing/providers"]) do %>
                       <%!-- Settings submenu items --%>
                       <div class="mt-1">
                         <.admin_nav_item
@@ -786,6 +787,16 @@ defmodule PhoenixKitWeb.Components.LayoutWrapper do
                             href={Routes.locale_aware_path(assigns, "/admin/settings/blogging")}
                             icon="document"
                             label={gettext("Blogging")}
+                            current_path={@current_path || ""}
+                            nested={true}
+                          />
+                        <% end %>
+
+                        <%= if PhoenixKit.Modules.Legal.enabled?() do %>
+                          <.admin_nav_item
+                            href={Routes.locale_aware_path(assigns, "/admin/settings/legal")}
+                            icon="legal"
+                            label={gettext("Legal")}
                             current_path={@current_path || ""}
                             nested={true}
                           />